远程调用

+++ auth-server/src/main/java/com/xinyuan/common/service/PageBean.java
@@ -61,7 +61,7 @@ public class PageBean implements Pageable {
      * 第二页所需要增加的数量
      */
     @Override
-    public int getOffset() {
+    public long getOffset() {
         return (getPagenumber() - 1) * getPagesize();
     }
 
@@ -105,4 +105,5 @@ public class PageBean implements Pageable {
     public boolean hasPrevious() {
         return false;
     }
+
 }
